Move cbc tests into read strategies tests
