+++ tests/test_read_strategies.py
@@ -1,7 +1,173 @@
+from io import StringIO
+from textwrap import dedent
+from typing import List
+
+from pytest import mark
+
 import pandas as pd
 from amply import Amply
 
-from otoole.read_strategies import ReadDatafile, ReadMemory
+from otoole import ReadDatafile, ReadMemory
+from otoole.results.results import ReadCbc
+
+
+class TestReadCbc:
+
+    test_data = [
+        (
+            dedent(
+                """Optimal - objective value 4483.96932237
+                             1 TotalDiscountedCost(SIMPLICITY,2015)                                                   187.01576                       0
+                             2 TotalDiscountedCost(SIMPLICITY,2016)                                                   183.30788                       0
+                             3 TotalDiscountedCost(SIMPLICITY,2017)                                                   181.05465                       0
+                             4 TotalDiscountedCost(SIMPLICITY,2018)                                                   218.08923                       0
+                             5 TotalDiscountedCost(SIMPLICITY,2019)                                                   193.85792                       0
+                             6 TotalDiscountedCost(SIMPLICITY,2020)                                                   233.79202                       0
+
+                         """
+            ),
+            pd.DataFrame(
+                data=[
+                    ["TotalDiscountedCost", "SIMPLICITY,2015", 187.01576],
+                    ["TotalDiscountedCost", "SIMPLICITY,2016", 183.30788],
+                    ["TotalDiscountedCost", "SIMPLICITY,2017", 181.05465],
+                    ["TotalDiscountedCost", "SIMPLICITY,2018", 218.08923],
+                    ["TotalDiscountedCost", "SIMPLICITY,2019", 193.85792],
+                    ["TotalDiscountedCost", "SIMPLICITY,2020", 233.79202],
+                ],
+                columns=["Variable", "Index", "Value"],
+            ),
+        )
+    ]
+
+    @mark.parametrize("cbc_input,expected", test_data, ids=["TotalDiscountedCost"])
+    def test_read_cbc_to_dataframe(self, cbc_input, expected):
+        cbc_reader = ReadCbc()
+        with StringIO(cbc_input) as file_buffer:
+            actual = cbc_reader._convert_cbc_to_dataframe(file_buffer)
+        pd.testing.assert_frame_equal(actual, expected)
+
+    test_data_2 = [
+        # First case
+        (
+            pd.DataFrame(
+                data=[
+                    ["TotalDiscountedCost", "SIMPLICITY,2015", 187.01576],
+                    ["TotalDiscountedCost", "SIMPLICITY,2016", 183.30788],
+                    ["TotalDiscountedCost", "SIMPLICITY,2017", 181.05465],
+                    ["TotalDiscountedCost", "SIMPLICITY,2018", 218.08923],
+                    ["TotalDiscountedCost", "SIMPLICITY,2019", 193.85792],
+                    ["TotalDiscountedCost", "SIMPLICITY,2020", 233.79202],
+                ],
+                columns=["Variable", "Index", "Value"],
+            ),
+            {},
+            (
+                {
+                    "TotalDiscountedCost": pd.DataFrame(
+                        data=[
+                            ["SIMPLICITY", 2015, 187.01576],
+                            ["SIMPLICITY", 2016, 183.30788],
+                            ["SIMPLICITY", 2017, 181.05465],
+                            ["SIMPLICITY", 2018, 218.08923],
+                            ["SIMPLICITY", 2019, 193.85792],
+                            ["SIMPLICITY", 2020, 233.79202],
+                        ],
+                        columns=["REGION", "YEAR", "VALUE"],
+                    ).set_index(["REGION", "YEAR"])
+                }
+            ),
+        ),
+        # Second case
+        (
+            pd.DataFrame(
+                data=[
+                    ["AnnualEmissions", "REGION,CO2,2017", 137958.8400384134],
+                    ["AnnualEmissions", "REGION,CO2,2018", 305945.3841061913],
+                    ["AnnualEmissions", "REGION,CO2,2019", 626159.9611543404],
+                ],
+                columns=["Variable", "Index", "Value"],
+            ),
+            {},
+            {
+                "AnnualEmissions": pd.DataFrame(
+                    data=[
+                        ["REGION", "CO2", 2017, 137958.8400384134],
+                        ["REGION", "CO2", 2018, 305945.3841061913],
+                        ["REGION", "CO2", 2019, 626159.9611543404],
+                    ],
+                    columns=["REGION", "EMISSION", "YEAR", "VALUE"],
+                ).set_index(["REGION", "EMISSION", "YEAR"])
+            },
+        ),
+    ]  # type: List
+
+    @mark.parametrize(
+        "results,cbc_input,expected",
+        test_data_2,
+        ids=["TotalDiscountedCost", "AnnualEmissions1"],
+    )
+    def test_convert_cbc_to_csv_long(self, results, cbc_input, expected):
+        cbc_reader = ReadCbc()
+        actual = cbc_reader._convert_dataframe_to_csv(results, cbc_input)
+        assert isinstance(actual, dict)
+        for name, df in actual.items():
+            pd.testing.assert_frame_equal(df, expected[name])
+
+    def test_convert_cbc_to_csv_short(self):
+        cbc_results = pd.DataFrame(
+            data=[
+                ["RateOfActivity", "SIMPLICITY,ID,GAS_EXTRACTION,1,2014", 1],
+                ["RateOfActivity", "SIMPLICITY,IN,GAS_EXTRACTION,1,2014", 1],
+                ["RateOfActivity", "SIMPLICITY,SD,GAS_EXTRACTION,1,2014", 1],
+                ["RateOfActivity", "SIMPLICITY,SN,GAS_EXTRACTION,1,2014", 1],
+                ["RateOfActivity", "SIMPLICITY,WD,GAS_EXTRACTION,1,2014", 1],
+                ["RateOfActivity", "SIMPLICITY,WN,GAS_EXTRACTION,1,2014", 1],
+                ["RateOfActivity", "SIMPLICITY,ID,DUMMY,1,2014", 1],
+                ["RateOfActivity", "SIMPLICITY,IN,DUMMY,1,2014", 1],
+                ["RateOfActivity", "SIMPLICITY,SD,DUMMY,1,2014", 1],
+                ["RateOfActivity", "SIMPLICITY,SN,DUMMY,1,2014", 1],
+                ["RateOfActivity", "SIMPLICITY,WD,DUMMY,1,2014", 1],
+                ["RateOfActivity", "SIMPLICITY,WN,DUMMY,1,2014", 1],
+            ],
+            columns=["Variable", "Index", "Value"],
+        )
+        input_data = {
+            "EmissionActivityRatio": pd.DataFrame(
+                data=[["SIMPLICITY", "GAS_EXTRACTION", "CO2", 1, 2014, 1.0]],
+                columns=[
+                    "REGION",
+                    "TECHNOLOGY",
+                    "EMISSION",
+                    "MODE_OF_OPERATION",
+                    "YEAR",
+                    "VALUE",
+                ],
+            ).set_index(
+                ["REGION", "TECHNOLOGY", "EMISSION", "MODE_OF_OPERATION", "YEAR"]
+            ),
+            "YearSplit": pd.DataFrame(
+                data=[
+                    ["ID", 2014, 0.1667],
+                    ["IN", 2014, 0.0833],
+                    ["SD", 2014, 0.1667],
+                    ["SN", 2014, 0.0833],
+                    ["WD", 2014, 0.3333],
+                    ["WN", 2014, 0.1667],
+                ],
+                columns=["TIMESLICE", "YEAR", "VALUE"],
+            ).set_index(["TIMESLICE", "YEAR"]),
+        }
+
+        expected = pd.DataFrame(
+            data=[["SIMPLICITY", "CO2", 2014, 1.0]],
+            columns=["REGION", "EMISSION", "YEAR", "VALUE"],
+        ).set_index(["REGION", "EMISSION", "YEAR"])
+
+        cbc_reader = ReadCbc()
+        actual = cbc_reader._convert_dataframe_to_csv(cbc_results, input_data)
+        assert isinstance(actual, dict)
+        pd.testing.assert_frame_equal(actual["AnnualEmissions"], expected)
 
 
 class TestReadMemoryStrategy:
